% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lacunarity.R
\name{lacunarity}
\alias{lacunarity}
\title{Lacunarity}
\usage{
lacunarity(
  x,
  r_vec = NULL,
  box = "SQUARE",
  plot = FALSE,
  save_plot = FALSE,
  progress = FALSE,
  ncores = 1L
)
}
\arguments{
\item{x}{SpatRaster or list of SpatRaster OR character; Raster image that can be loaded using \code{\link[terra]{rast}}.
x can also be character, refering to a folder. In that case all files in the folder with the ending ".tif" will be used}

\item{r_vec}{integer vector (optional); Vector of box diameter values. Every r must be odd and >2. It is recommended
that no r value is greated than half of the shorter dimension of x. If r_vec is NULL (default), r_vec will be generated automatically}

\item{box}{character; Either SQUARE for a square neighborhood window, or CIRCLE for a round neighborhood window}

\item{plot}{logical; Should the summary Lacunarity figure, showing Lacunarity of all SpatRasters be printed?}

\item{save_plot}{FALSE or folder path; If not FALSE, a folder path to save Lacunarity plots (see details)}

\item{progress}{logical; Show progress bar?}

\item{ncores}{numeric; The number of cores to use}
}
\value{
\code{\link[tibble]{tibble}} containing all Lacunarity values: \code{name} and \code{i} refer to the name and index of the raster input.
\code{x} is the box diameter (a 5X5 box has a diameter of 5 with a radius of 2). \code{lac} indicates the Lacunarity value.
}
\description{
Computes the Lacunarity of a Spatial Raster object or a list of Spatial Raster objects.
}
\details{
\code{lacunarity} is based on the algorithm for binary images provided by \href{https://doi.org/10.1007/BF00125351}{Plotnick et al. 1993}.
\href{https://doi.org/10.1016/j.ecocom.2011.01.001}{Hoechstetter et al. 2011} further applyed this algorithm on continuous raster images.
If \code{x} is a binary raster (e.g. Land Use) Lacunarity is beening calculated using the Plotnicks algorithm.
}
\examples{
# Create a SpatRast as input
mat_sample <- matrix(data = c(
   1,1,0,1,1,1,0,1,0,1,1,0,
   0,0,0,0,0,1,0,0,0,1,1,1,
   0,1,0,1,1,1,1,1,0,1,1,0,
   1,0,1,1,1,0,0,0,0,0,0,0,
   1,1,0,1,0,1,0,0,1,1,0,0,
   0,1,0,1,1,0,0,1,0,0,1,0,
   0,0,0,0,0,1,1,1,1,1,1,1,
   0,1,1,0,0,0,1,1,1,1,0,0,
   0,1,1,1,0,1,1,0,1,0,0,1,
   0,1,0,0,0,0,0,0,0,1,1,1,
   0,1,0,1,1,1,0,1,1,0,1,0,
   0,1,0,0,0,1,0,1,1,1,0,1
   ), nrow = 12, ncol = 12, byrow = TRUE
)
x <- terra::rast(mat_sample)

lacunarity(x)

}
