\name{spatLac-package}
\alias{spatLac-package}
\alias{spatLac}
\docType{package}
\title{\packageTitle{spatLac}}
\description{\packageDescription{spatLac}}
\author{
\strong{Maintainer}: Sebastian Brinkmann \email{sebastian.brinkmann@fau.de} (\href{https://orcid.org/0000-0001-9835-7347}{ORCID})

Authors:
\itemize{
  \item Sebastian Brinkmann (\href{https://orcid.org/0000-0001-9835-7347}{ORCID})
  \item S.M. Labib (\href{https://orcid.org/0000-0002-4127-2075}{ORCID})
}

}
\keyword{package}
