% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lacunarity.R
\name{lacunarity}
\alias{lacunarity}
\title{Lacunarity}
\usage{
lacunarity(
  r,
  box = "SQUARE",
  plot = FALSE,
  save_plot = FALSE,
  progress = FALSE,
  ncores = 1L
)
}
\arguments{
\item{r}{SpatRaster or list of SpatRaster objects; Raster image can be loaded with \code{\link[terra]{rast}}}

\item{box}{character; Either SQUARE for a square neighborhood window, or CIRCLE for a round neighborhood window}

\item{plot}{logical; Should the summary Lacunarity figure, showing Lacunarity of all SpatRasters be printed?}

\item{save_plot}{FALSE or folder path; If not FALSE, a folder path to save Lacunarity plots (see details)}

\item{progress}{logical; Show progress bar?}

\item{ncores}{numeric; The number of cores to use}
}
\value{
\code{\link[dplyr]{tibble}} containing all Lacunarity values
}
\description{
Computes the Lacunarity of a Spatial Raster object or a list of Spatial Raster objects.
}
\details{
\code{lacunarity} is based on the algorithm for binary images provided by \href{https://doi.org/10.1007/BF00125351}{Plotnick et al. 1993}.
\href{https://doi.org/10.1016/j.ecocom.2011.01.001}{Hoechstetter et al. 2011} further applyed this algorithm on continuous raster images.
If \code{r} is a binary raster (e.g. Land Use) Lacunarity is beening calculated using the Plotnicks algorithm.
}
\examples{
# Create a SpatRast as input
mat_sample <- matrix(data = c(
   1,1,0,1,1,1,0,1,0,1,1,0,
   0,0,0,0,0,1,0,0,0,1,1,1,
   0,1,0,1,1,1,1,1,0,1,1,0,
   1,0,1,1,1,0,0,0,0,0,0,0,
   1,1,0,1,0,1,0,0,1,1,0,0,
   0,1,0,1,1,0,0,1,0,0,1,0,
   0,0,0,0,0,1,1,1,1,1,1,1,
   0,1,1,0,0,0,1,1,1,1,0,0,
   0,1,1,1,0,1,1,0,1,0,0,1,
   0,1,0,0,0,0,0,0,0,1,1,1,
   0,1,0,1,1,1,0,1,1,0,1,0,
   0,1,0,0,0,1,0,1,1,1,0,1
   ), nrow = 12, ncol = 12, byrow = TRUE
)
r <- terra::rast(mat_sample)

lacunarity(r)

}
